+++ src/parser/Node.cpp
@@ -6,7 +6,7 @@
 
 std::vector<std::string> Node::stackTrace;
 
-std::string parseString(const std::string &source)
+std::string Node::parseString(const std::string &source)
 {
     std::string ret;
     size_t length = source.size() - 1;
@@ -109,7 +109,6 @@ void Node::addChild(Node &arg)
 
 object::objectPtr Node::evaluate(stack &st) const
 {
-    //console::debug(toString());
     switch (_token)
     {
     case COMPOUND_STATEMENT:
@@ -119,7 +118,7 @@ object::objectPtr Node::evaluate(stack &st) const
         try
         {
             for (; i < _children.size(); i++)
-                _children[i].evaluate(localStack);
+                _children[i].evaluateVoid(localStack);
         }
         catch (objectException &e)
         {
@@ -137,30 +136,42 @@ object::objectPtr Node::evaluate(stack &st) const
     case IF:
         assert(_children.size() == 2);
         if (_children[0].evaluateBoolean(st))
-            _children[1].evaluate(st);
+            _children[1].evaluateVoid(st);
         return nullptr;
 
     case IF_ELSE:
         assert(_children.size() == 3);
         if (_children[0].evaluateBoolean(st))
-            _children[1].evaluate(st);
+            _children[1].evaluateVoid(st);
         else
-            _children[2].evaluate(st);
+            _children[2].evaluateVoid(st);
         return nullptr;
 
     case WHILE:
         assert(_children.size() == 2);
-        while (_children[0].evaluateBoolean(st))
-            _children[1].evaluate(st);
+        try
+        {
+            while (_children[0].evaluateBoolean(st))
+                _children[1].evaluateVoid(st);
+        }
+        catch (const breakType &e)
+        {
+        }
         return nullptr;
 
     case FOR:
         assert(_children.size() == 4);
-        _children[0].evaluate(st);
-        while (_children[1].evaluateBoolean(st))
+        _children[0].evaluateVoid(st);
+        try
         {
-            _children[3].evaluate(st);
-            _children[2].evaluate(st);
+            while (_children[1].evaluateBoolean(st))
+            {
+                _children[3].evaluateVoid(st);
+                _children[2].evaluateVoid(st);
+            }
+        }
+        catch (const breakType &e)
+        {
         }
         return nullptr;
 
@@ -341,6 +352,26 @@ object::objectPtr Node::evaluate(stack &st) const
             return st.insert(static_cast<name>(_children[0].evaluate(st)->getConverted<std::string>()), a);
     }
 
+    case DELETE_:
+        assert(_children.size() == 1);
+        if (_children[0]._token == IDENTIFIER)
+        {
+            assert(!_children[0]._text.isEmpty());
+            auto a = st[_children[0]._text];
+            st.erase(_children[0]._text);
+            return a;
+        }
+        else if (_children[0]._token == DOT && _children[0]._children[1]._token == IDENTIFIER)
+        {
+            assert(_children[0]._children.size() == 2);
+            auto expr = _children[0]._children[0].evaluate(st);
+            auto a = (*expr)[_children[0]._children[1]._text];
+            expr->removeProperty(_children[0]._children[1]._text);
+            return a;
+        }
+        throw std::runtime_error("left side of delete is not identifier or dot operator");
+        break;
+
     case JSON:
     {
         stack localJsonStack(&st);
@@ -383,26 +414,29 @@ object::objectPtr Node::evaluate(stack &st) const
         assert(_children.size() == 1);
         throw objectException(_children[0].evaluate(st));
 
+    case BREAK:
+        throw breakType();
+
     case TRY:
     {
         assert(_children.size() == 2);
         try
         {
-            _children[0].evaluate(st);
+            _children[0].evaluateVoid(st);
         }
         catch (objectException &e)
         {
             stackTrace.clear();
             stack localStack(&st);
             localStack.insert("exception"_n, e.getPtr());
-            _children[1].evaluate(localStack);
+            _children[1].evaluateVoid(localStack);
         }
         catch (std::exception &e)
         {
             stackTrace.clear();
             stack localStack(&st);
             localStack.insert("exception"_n, object::makeObject((std::string)e.what()));
-            _children[1].evaluate(localStack);
+            _children[1].evaluateVoid(localStack);
         }
         return nullptr;
     }
@@ -431,9 +465,9 @@ object::objectPtr Node::evaluate(stack &st) const
         auto b = _children[1].evaluate(st);
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<number>(a->get<const number>() + b->getConverted<number>()));
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<std::string>(a->get<const std::string>() + b->getConverted<std::string>()));
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
         {
             // TODO: in place
             object::arrayType arr = a->get<const object::arrayType>();
@@ -451,7 +485,7 @@ object::objectPtr Node::evaluate(stack &st) const
         auto b = _children[1].evaluate(st);
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<number>(a->get<const number>() * b->getConverted<number>()));
-        if (a->isOfType<std::string>() && b->isConvertible<number>())
+        else if (a->isOfType<std::string>() && b->isConvertible<number>())
         {
             const std::string &str = a->get<const std::string>();
             int i = static_cast<int>(b->getConverted<number>());
@@ -461,7 +495,7 @@ object::objectPtr Node::evaluate(stack &st) const
                 res += str;
             return object::makeObject(res);
         }
-        if (a->isOfType<object::arrayType>() && b->isConvertible<number>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<number>())
         {
             const object::arrayType &arr = a->get<const object::arrayType>();
             int i = static_cast<int>(b->getConverted<number>());
@@ -509,9 +543,9 @@ object::objectPtr Node::evaluate(stack &st) const
         assert(_children.size() == 2);
         auto a = _children[0].evaluate(st);
         object::objectPtr b;
-        if (a->isOfType<bool>())
+        if (a->isConvertible<bool>())
         {
-            if (a->get<const bool>() == false)
+            if (!a->getConverted<bool>())
                 return object::makeObject(false);
             b = _children[1].evaluate(st);
             if (b->isConvertible<bool>())
@@ -526,10 +560,18 @@ object::objectPtr Node::evaluate(stack &st) const
     {
         assert(_children.size() == 2);
         auto a = _children[0].evaluate(st);
-        auto b = _children[1].evaluate(st);
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
-            return object::makeObject(static_cast<bool>(a->get<const bool>() || b->getConverted<bool>()));
-        return (*(*a)["||"_n])(a, {b}, &st);
+        object::objectPtr b;
+        if (a->isConvertible<bool>())
+        {
+            if (a->getConverted<bool>())
+                return a;
+            b = _children[1].evaluate(st);
+            if (b->isConvertible<bool>())
+                return b;
+        }
+        else
+            b = _children[1].evaluate(st);
+        return (*(*a)["&&"_n])(a, {b}, &st);
     }
 
     case BITWISE_AND:
@@ -580,13 +622,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() == b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() == b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() == b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() == b->getConverted<bool>()));
 
         return (*(*a)["=="_n])(a, {b}, &st);
@@ -600,13 +642,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() != b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() != b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() != b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() != b->getConverted<bool>()));
 
         return (*(*a)["!="_n])(a, {b}, &st);
@@ -620,13 +662,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() < b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() < b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() < b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() < b->getConverted<bool>()));
 
         return (*(*a)["<"_n])(a, {b}, &st);
@@ -640,13 +682,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() > b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() > b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() > b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() > b->getConverted<bool>()));
 
         return (*(*a)[">"_n])(a, {b}, &st);
@@ -660,13 +702,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() <= b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() <= b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() <= b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() <= b->getConverted<bool>()));
 
         return (*(*a)["<="_n])(a, {b}, &st);
@@ -680,13 +722,13 @@ object::objectPtr Node::evaluate(stack &st) const
         if (a->isOfType<number>() && b->isConvertible<number>())
             return object::makeObject(static_cast<bool>(a->get<const number>() >= b->getConverted<number>()));
 
-        if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
             return object::makeObject(static_cast<bool>(a->get<const std::string>() >= b->getConverted<std::string>()));
 
-        if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
             return object::makeObject(static_cast<bool>(a->get<const object::arrayType>() >= b->getConverted<object::arrayType>()));
 
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
             return object::makeObject(static_cast<bool>(a->get<const bool>() >= b->getConverted<bool>()));
 
         return (*(*a)[">="_n])(a, {b}, &st);
@@ -755,6 +797,672 @@ object::objectPtr Node::evaluate(stack &st) const
     return object::makeEmptyObject();
 }
 
+void Node::evaluateVoid(stack &st) const
+{
+    switch (_token)
+    {
+    case COMPOUND_STATEMENT:
+    {
+        stack localStack(&st);
+        size_t i = 0;
+        try
+        {
+            for (; i < _children.size(); i++)
+                _children[i].evaluateVoid(localStack);
+        }
+        catch (objectException &e)
+        {
+            _children[i].exception();
+            throw e;
+        }
+        catch (std::exception &e)
+        {
+            _children[i].exception();
+            throw e;
+        }
+        return;
+    }
+
+    case IF:
+        assert(_children.size() == 2);
+        if (_children[0].evaluateBoolean(st))
+            _children[1].evaluateVoid(st);
+        return;
+
+    case IF_ELSE:
+        assert(_children.size() == 3);
+        if (_children[0].evaluateBoolean(st))
+            _children[1].evaluateVoid(st);
+        else
+            _children[2].evaluateVoid(st);
+        return;
+
+    case WHILE:
+        assert(_children.size() == 2);
+        try
+        {
+            while (_children[0].evaluateBoolean(st))
+                _children[1].evaluateVoid(st);
+        }
+        catch (const breakType &e)
+        {
+        }
+        return;
+
+    case FOR:
+        assert(_children.size() == 4);
+        _children[0].evaluateVoid(st);
+        try
+        {
+            while (_children[1].evaluateBoolean(st))
+            {
+                _children[3].evaluateVoid(st);
+                _children[2].evaluateVoid(st);
+            }
+        }
+        catch (const breakType &e)
+        {
+        }
+        return;
+
+    case STATEMENT:
+        assert(_children.size() == 1);
+        _children[0].evaluateVoid(st);
+        return;
+
+    case CALL_OPERATOR:
+    {
+        assert(_children.size() >= 1);
+        auto a = _children[0].evaluate(st);
+        object::arrayType args(_children.size() - 1);
+        for (size_t i = 1; i < _children.size(); i++)
+            args[i - 1] = _children[i].evaluate(st);
+
+        try
+        {
+            (*a)(a, std::move(args), &st);
+            return;
+        }
+        catch (objectException &e)
+        {
+            _children[0].exception(_children[0].toName());
+            throw e;
+        }
+        catch (std::exception &e)
+        {
+            _children[0].exception(_children[0].toName());
+            throw e;
+        }
+    }
+
+    case READ_OPERATOR:
+    {
+        assert(_children.size() >= 1);
+        static name readOperator = "readOperator"_n;
+        auto a = _children[0].evaluate(st);
+        object::arrayType args(_children.size() - 1);
+        for (size_t i = 1; i < _children.size(); i++)
+            args[i - 1] = _children[i].evaluate(st);
+
+        try
+        {
+            (*(*a)[readOperator])(a, std::move(args), &st);
+            return;
+        }
+        catch (objectException &e)
+        {
+            _children[0].exception(_children[0].toName() + ".readOperator");
+            throw e;
+        }
+        catch (std::exception &e)
+        {
+            _children[0].exception(_children[0].toName() + ".readOperator");
+            throw e;
+        }
+    }
+
+    case METHOD_CALL_OPERATOR:
+    {
+        assert(_children.size() >= 2);
+        assert(_children[1]._token == IDENTIFIER);
+        auto a = _children[0].evaluate(st);
+        auto b = (*a)[_children[1]._text];
+        object::arrayType args(_children.size() - 2);
+        for (size_t i = 2; i < _children.size(); i++)
+            args[i - 2] = _children[i].evaluate(st);
+
+        try
+        {
+            (*b)(a, std::move(args), &st);
+            return;
+        }
+        catch (objectException &e)
+        {
+            _children[0].exception(_children[0].toName() + "." + _children[1].toName());
+            throw e;
+        }
+        catch (std::exception &e)
+        {
+            _children[0].exception(_children[0].toName() + "." + _children[1].toName());
+            throw e;
+        }
+    }
+
+    case AWAIT:
+    {
+        assert(_children.size() == 1);
+        auto a = _children[0].evaluate(st);
+        auto b = (*a)["await"_n];
+        (*b)(a, {}, &st);
+        return;
+    }
+
+    case ARRAY_LITERAL:
+    {
+        for (size_t i = 0; i < _children.size(); i++)
+            _children[i].evaluateVoid(st);
+        return;
+    }
+
+    case DOT:
+    {
+        assert(_children.size() == 2);
+        assert(_children[1]._token == IDENTIFIER);
+        (*(_children[0].evaluate(st)))[_children[1]._text];
+        return;
+    }
+
+    case LET:
+        assert(_children.size() == 0);
+        st.insert(_text, object::makeEmptyObject());
+        return;
+
+    case ASSIGNMENT:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        if (a->isConst())
+            throw std::runtime_error("tried to modify constant value");
+        if (data->numeric >= 100)
+        {
+            try
+            {
+                a->get<number>() = _children[1].evaluateNumber(st);
+                return;
+            }
+            catch (...)
+            {
+                console::warn("bad number optimization");
+                data->numeric -= 50;
+            }
+        }
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isOfType<number>())
+            data->numeric = std::min(data->numeric + 1, 1000);
+        *a = *b;
+        a->setConst(false);
+        return;
+    }
+
+    case INIT_ASSIGNMENT:
+        assert(_children.size() == 2);
+        if (_children[0]._token == LET)
+        {
+            st.insert(_children[0]._text, _children[1].evaluate(st));
+            return;
+        }
+        else if (_children[0]._token == IDENTIFIER)
+        {
+            assert(!_children[0]._text.isEmpty());
+            auto &a = st[_children[0]._text];
+            if (a->isConst())
+                throw std::runtime_error("tried to modify constant value");
+            a = _children[1].evaluate(st);
+            return;
+        }
+        else if (_children[0]._token == DOT && _children[0]._children[1]._token == IDENTIFIER)
+        {
+            assert(_children[0]._children.size() == 2);
+            auto &a = (*(_children[0]._children[0].evaluate(st)))[_children[0]._children[1]._text];
+            if (a->isConst())
+                throw std::runtime_error("tried to modify constant value");
+            a = _children[1].evaluate(st);
+            return;
+        }
+        throw std::runtime_error("left side of init_assignment is not identifier or dot operator");
+        break;
+
+    case JSON_ASSIGNMENT:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[1].evaluate(st);
+        if (a->isConst())
+        {
+            auto b = object::makeEmptyObject();
+            *b = *a;
+            b->setConst(false);
+            a = b;
+        }
+        if (_children[0]._token == IDENTIFIER)
+            st.insert(_children[0]._text, a);
+        else
+            st.insert(static_cast<name>(_children[0].evaluate(st)->getConverted<std::string>()), a);
+        return;
+    }
+
+    case DELETE_:
+        assert(_children.size() == 1);
+        if (_children[0]._token == IDENTIFIER)
+        {
+            assert(!_children[0]._text.isEmpty());
+            auto a = st[_children[0]._text];
+            st.erase(_children[0]._text);
+            return;
+        }
+        else if (_children[0]._token == DOT && _children[0]._children[1]._token == IDENTIFIER)
+        {
+            assert(_children[0]._children.size() == 2);
+            auto expr = _children[0]._children[0].evaluate(st);
+            auto a = (*expr)[_children[0]._children[1]._text];
+            expr->removeProperty(_children[0]._children[1]._text);
+            return;
+        }
+        throw std::runtime_error("left side of delete is not identifier or dot operator");
+        break;
+
+    case JSON:
+    {
+        stack localJsonStack(&st);
+        for (auto &child : _children)
+            child.evaluate(localJsonStack);
+
+        auto obj = object::makeEmptyObject();
+        for (auto x : localJsonStack)
+            obj->addProperty(x.first, x.second);
+        return;
+    }
+
+    case CLASS:
+    {
+        stack localJsonStack(&st);
+        for (auto &child : _children)
+            child.evaluate(localJsonStack);
+
+        auto obj = object::makeEmptyObject();
+        for (auto x : localJsonStack)
+            obj->addProperty(x.first, x.second);
+
+        auto Class = st["Class"_n];
+        if (!_children.empty() && _children.back()._token == IDENTIFIER)
+        {
+            auto &prototype = (*obj)[name::prototype];
+            prototype = (*_children.back().evaluate(st))["classPrototype"_n];
+            (*prototype)["super"_n] = (*prototype)["constructor"_n];
+        }
+        obj = (*Class)(Class, {obj}, &st);
+        st.insert(_text, obj);
+        return;
+    }
+
+    case RETURN:
+        assert(_children.size() == 1);
+        throw _children[0].evaluate(st);
+
+    case THROW:
+        assert(_children.size() == 1);
+        throw objectException(_children[0].evaluate(st));
+
+    case BREAK:
+        throw breakType();
+
+    case TRY:
+    {
+        assert(_children.size() == 2);
+        try
+        {
+            _children[0].evaluateVoid(st);
+        }
+        catch (objectException &e)
+        {
+            stackTrace.clear();
+            stack localStack(&st);
+            localStack.insert("exception"_n, e.getPtr());
+            _children[1].evaluateVoid(localStack);
+        }
+        catch (std::exception &e)
+        {
+            stackTrace.clear();
+            stack localStack(&st);
+            localStack.insert("exception"_n, object::makeObject((std::string)e.what()));
+            _children[1].evaluateVoid(localStack);
+        }
+        return;
+    }
+
+    case USER_OPERATOR:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        auto op = st[_text];
+        (*op)(op, {a, b}, &st);
+        return;
+    }
+
+    case INSTANCEOF:
+        return;
+
+    case ADDITION:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        (*(*a)["+"_n])(a, {b}, &st);
+        return;
+    }
+
+    case MULTIPLICATION:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<number>())
+            return;
+        (*(*a)["*"_n])(a, {b}, &st);
+        return;
+    }
+
+    case SUBTRACTION:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["-"_n])(a, {b}, &st);
+        return;
+    }
+
+    case DIVISION:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["/"_n])(a, {b}, &st);
+        return;
+    }
+
+    case MODULUS:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["%"_n])(a, {b}, &st);
+        return;
+    }
+
+    case AND:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        object::objectPtr b;
+        if (a->isConvertible<bool>())
+        {
+            if (!a->getConverted<bool>())
+                return;
+            b = _children[1].evaluate(st);
+            if (b->isConvertible<bool>())
+                return;
+        }
+        else
+            b = _children[1].evaluate(st);
+        (*(*a)["&&"_n])(a, {b}, &st);
+        return;
+    }
+
+    case OR:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        object::objectPtr b;
+        if (a->isConvertible<bool>())
+        {
+            if (a->getConverted<bool>())
+                return;
+            b = _children[1].evaluate(st);
+            if (b->isConvertible<bool>())
+                return;
+        }
+        else
+            b = _children[1].evaluate(st);
+        (*(*a)["||"_n])(a, {b}, &st);
+        return;
+    }
+
+    case BITWISE_AND:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["&"_n])(a, {b}, &st);
+        return;
+    }
+
+    case BITWISE_OR:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["|"_n])(a, {b}, &st);
+        return;
+    }
+
+    case SHIFT_LEFT:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)["<<"_n])(a, {b}, &st);
+        return;
+    }
+
+    case SHIFT_RIGHT:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        (*(*a)[">>"_n])(a, {b}, &st);
+        return;
+    }
+
+    case EQUAL:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)["=="_n])(a, {b}, &st);
+        return;
+    }
+
+    case NOT_EQUAL:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)["!="_n])(a, {b}, &st);
+        return;
+    }
+
+    case LESS:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)["<"_n])(a, {b}, &st);
+        return;
+    }
+
+    case GREATER:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)[">"_n])(a, {b}, &st);
+        return;
+    }
+
+    case LESS_EQUAL:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)["<="_n])(a, {b}, &st);
+        return;
+    }
+
+    case GREATER_EQUAL:
+    {
+        assert(_children.size() == 2);
+        auto a = _children[0].evaluate(st);
+        auto b = _children[1].evaluate(st);
+        if (a->isOfType<number>() && b->isConvertible<number>())
+            return;
+        else if (a->isOfType<std::string>() && b->isConvertible<std::string>())
+            return;
+        else if (a->isOfType<object::arrayType>() && b->isConvertible<object::arrayType>())
+            return;
+        else if (a->isOfType<bool>() && b->isConvertible<bool>())
+            return;
+        (*(*a)[">="_n])(a, {b}, &st);
+        return;
+    }
+
+    case INCREMENT:
+    {
+        assert(_children.size() == 1);
+        auto a = _children[0].evaluate(st);
+        if (a->isOfType<number>())
+        {
+            a->get<number>()++;
+            return;
+        }
+        (*(*a)["++"_n])(a, {}, &st);
+        return;
+    }
+
+    case DECREMENT:
+    {
+        assert(_children.size() == 1);
+        auto a = _children[0].evaluate(st);
+        if (a->isOfType<number>())
+        {
+            a->get<number>()--;
+            return;
+        }
+        (*(*a)["--"_n])(a, {}, &st);
+        return;
+    }
+
+    case NOT:
+    {
+        assert(_children.size() == 1);
+        auto a = _children[0].evaluate(st);
+        if (a->isConvertible<bool>())
+            return;
+        (*(*a)["!"_n])(a, {}, &st);
+        return;
+    }
+
+    case COMPLEMENT:
+    {
+        assert(_children.size() == 1);
+        auto a = _children[0].evaluate(st);
+        if (a->isOfType<number>())
+            return;
+        (*(*a)["~"_n])(a, {}, &st);
+        return;
+    }
+
+    case IDENTIFIER:
+        return;
+
+    case FUNCTION:
+        if (!_text.isEmpty())
+            st.insert(_text, data->value);
+        data->value->captureStack(st.flatCopy());
+    case NUMBER_LITERAL:
+    case STRING_LITERAL:
+        return;
+
+    default:
+        for (auto &child : _children)
+            console::warn(LLgetSymbol(child._token));
+        break;
+    }
+}
+
 number Node::evaluateNumber(stack &st) const
 {
     switch (_token)
@@ -763,7 +1471,7 @@ number Node::evaluateNumber(stack &st) const
     //case CALL_OPERATOR:
     //case METHOD_CALL_OPERATOR:
     // TODO
-    // sin, cos, sqrt etc. 
+    // sin, cos, sqrt etc.
     case ADDITION:
     {
         assert(_children.size() == 2);
@@ -836,7 +1544,7 @@ number Node::evaluateNumber(stack &st) const
         return static_cast<number>(static_cast<int>(a) >> static_cast<int>(b));
     }
 
-    /*case INCREMENT:
+        /*case INCREMENT:
     {
         assert(_children.size() == 1);
         auto a = _children[0].evaluate(st);
@@ -883,9 +1591,9 @@ bool Node::evaluateBoolean(stack &st) const
         assert(_children.size() == 2);
         auto a = _children[0].evaluate(st);
         object::objectPtr b;
-        if (a->isOfType<bool>())
+        if (a->isConvertible<bool>())
         {
-            if (a->get<const bool>() == false)
+            if (!a->getConverted<bool>())
                 return false;
             b = _children[1].evaluate(st);
             if (b->isConvertible<bool>())
@@ -893,16 +1601,24 @@ bool Node::evaluateBoolean(stack &st) const
         }
         else
             b = _children[1].evaluate(st);
-        return (*(*a)["&&"_n])(a, {b}, &st);
+        return (*(*a)["&&"_n])(a, {b}, &st)->getConverted<bool>();
     }
 
     case OR:
     {
         assert(_children.size() == 2);
         auto a = _children[0].evaluate(st);
-        auto b = _children[1].evaluate(st);
-        if (a->isOfType<bool>() && b->isConvertible<bool>())
-            return static_cast<bool>(a->get<const bool>() || b->getConverted<bool>());
+        object::objectPtr b;
+        if (a->isConvertible<bool>())
+        {
+            if (a->getConverted<bool>())
+                return true;
+            b = _children[1].evaluate(st);
+            if (b->isConvertible<bool>())
+                return b->getConverted<bool>();
+        }
+        else
+            b = _children[1].evaluate(st);
         return (*(*a)["||"_n])(a, {b}, &st)->getConverted<bool>();
     }
 
@@ -1026,7 +1742,6 @@ bool Node::evaluateBoolean(stack &st) const
         return (*(*a)[">="_n])(a, {b}, &st)->getConverted<bool>();
     }
 
-
     case NOT:
     {
         assert(_children.size() == 1);
@@ -1035,7 +1750,6 @@ bool Node::evaluateBoolean(stack &st) const
             return !a->getConverted<bool>();
         return (*(*a)["!"_n])(a, {}, &st)->getConverted<bool>();
     }
-
     }
     return evaluate(st)->getConverted<bool>();
 }
@@ -1140,6 +1854,8 @@ std::string Node::toString() const
         return "return " + _children[0].toString();
     case THROW:
         return "return " + _children[0].toString();
+    case BREAK:
+        throw breakType();
     case TRY:
     {
         std::string r = "try ";